Move all tests for monkey_island/cc/services/reporting/report.py to one file
